Scrub and document picoquic.h
