+++ picohttp/demoserver.h
@@ -156,8 +156,12 @@ int picoquic_demo_server_callback(picoquic_cnx_t* cnx,
 
 size_t picoquic_demo_server_callback_select_alpn(picoquic_quic_t* quic, ptls_iovec_t* list, size_t count);
 
-#endif /* DEMO_SERVER_H */
-
 int demo_server_is_path_sane(const uint8_t* path, size_t path_length);
 
 int demo_server_try_file_path(const uint8_t* path, size_t path_length, size_t* echo_size, FILE** pF, char const* web_folder);
+
+/* For building a basic HTTP 0.9 test server */
+int http0dot9_get(uint8_t* command, size_t command_length,
+    uint8_t* response, size_t response_max, size_t* response_length);
+
+#endif /* DEMO_SERVER_H */
